Исправляет ошибки в кнопке - больше событий этого автора
