+++ src/partials/modal.html
@@ -42,7 +42,7 @@
         </ul>
       </div>
     </div>
-    <button class="modal__btn modal__btn-more" type="button">More from this author</button>
+    <button class="modal__btn modal__btn-more" type="button" data>More from this author</button>
 
   </div>
 </div>
