+++ src/sass/_bottom-part.scss
@@ -1,6 +1,6 @@
 .bottom-part_section {
   @media screen and(min-width: 768px) {
-    min-height: 2582px;
+    max-height: 2582px;
 
     // min-height: 2454px;
     // margin-left: auto;
@@ -16,13 +16,20 @@
   }
 
   @media screen and(min-width: 1280px) {
-    min-height: 1930px;
+    max-height: 1930px;
   }
 }
 
 .bottom-part {
+  @media screen and(max-width: 767px) {
+    background-image: url('/images/bg-img/dg-decoration/mobile/body-bg-mobile.svg');
+    background-size: auto;
+    background-position: 0 0;
+  }
+
   @media screen and(min-width: 768px) and (max-width: 1279px) {
-    background-image: url('/images/bg-img/dg-decoration/desktop/line_1.svg'),
+    background-image: url('/images/bg-img/dg-decoration/desktop/body-bg-desktop.svg'),
+      url('/images/bg-img/dg-decoration/desktop/line_1.svg'),
       url('/images/bg-img/dg-decoration/ellipse_5.svg'),
       url('/images/bg-img/dg-decoration/ellipse_2.svg'),
       url('/images/bg-img/dg-decoration/desktop/line_2.svg'),
@@ -36,20 +43,21 @@
       url('/images/bg-img/dg-decoration/desktop/line_6.svg'),
       url('/images/bg-img/dg-decoration/ellipse_9.svg');
 
-    background-size: 1px 220px, 150px 150px, 80px 80px, 1px 2288px, 237px 237px, 126px 126px,
+    background-size: auto, 1px 220px, 150px 150px, 80px 80px, 1px 2288px, 237px 237px, 126px 126px,
       80px 80px, 1px 1400px, 237px 237px, 126px 126px, 80px 80px, 1px 1295px, 237px 237px;
 
-    background-position: calc(50% - 298px) top, calc(50% - 298px) 200px, calc(50% - 298px) 235px,
-      calc(50% - 298px) 333px, calc(50% - 238px) 930px, calc(50% - 298px) 2587px,
-      calc(50% - 298px) 2617px, calc(50% + 298px) top, calc(50% + 238px) 350px,
-      calc(50% + 298px) 1400px, calc(50% + 298px) 1425px, calc(50% + 298px) 1525px,
-      calc(50% + 238px) 1800px;
+    background-position: 0 0, calc(50% - 298px) top, calc(50% - 298px) 200px,
+      calc(50% - 298px) 235px, calc(50% - 298px) 333px, calc(50% - 238px) 930px,
+      calc(50% - 298px) 2587px, calc(50% - 298px) 2617px, calc(50% + 298px) top,
+      calc(50% + 238px) 350px, calc(50% + 298px) 1400px, calc(50% + 298px) 1425px,
+      calc(50% + 298px) 1525px, calc(50% + 238px) 1800px;
 
     background-repeat: no-repeat;
   }
 
   @media screen and(min-width: 1280px) {
-    background-image: url('/images/bg-img/dg-decoration/line_1.svg'),
+    background-image: url('../images/bg-img/dg-decoration/body-bg-tablet.svg'),
+      url('/images/bg-img/dg-decoration/line_1.svg'),
       url('/images/bg-img/dg-decoration/ellipse_5.svg'),
       url('/images/bg-img/dg-decoration/ellipse_2.svg'),
       url('/images/bg-img/dg-decoration/line_3.svg'),
@@ -63,14 +71,14 @@
       url('/images/bg-img/dg-decoration/line_2_1.svg'),
       url('/images/bg-img/dg-decoration/ellipse_9.svg');
 
-    background-size: 1px 405px, 305px 305px, 150px 150px, 1px 1290px, 334px 334px, 216px 216px,
+    background-size: auto, 1px 405px, 305px 305px, 150px 150px, 1px 1290px, 334px 334px, 216px 216px,
       136px 136px, 1px 1267px, 334px 334px, 216px 216px, 136px 136px, 1px 993px, 334px 334px;
 
-    background-position: calc(50% - 525px) top, calc(50% - 525px) 365px, calc(50% - 525px) 445px,
-      calc(50% - 525px) 632px, calc(50% - 450px) 930px, calc(50% - 525px) 1870px,
-      calc(50% - 525px) 1925px, calc(50% + 525px) -195px, calc(50% + 455px) 44px,
-      calc(50% + 525px) 880px, calc(50% + 525px) 920px, calc(50% + 525px) 1100px,
-      calc(50% + 455px) 1500px;
+    background-position: 0 0, calc(50% - 525px) top, calc(50% - 525px) 365px,
+      calc(50% - 525px) 445px, calc(50% - 525px) 632px, calc(50% - 450px) 930px,
+      calc(50% - 525px) 1870px, calc(50% - 525px) 1925px, calc(50% + 525px) -195px,
+      calc(50% + 455px) 44px, calc(50% + 525px) 880px, calc(50% + 525px) 920px,
+      calc(50% + 525px) 1100px, calc(50% + 455px) 1500px;
 
     background-repeat: no-repeat;
   }
