+++ src/templates/modalContent.hbs
@@ -57,5 +57,5 @@
     </div>
 </div>
 
-<button class="modal__btn modal__btn-more" type="button" data-idattraction={{idAttraction}}>More from this
+<button class="modal__btn modal__btn-more" type="button" data-action="load" data-idattraction={{idAttraction}}>More from this
     author</button>
